refactor: move createTemplate graphql resolver to a separate file
