+++ templates/graphql/index.js
@@ -13,24 +13,13 @@ const {
   validateFiles,
 } = require("./utils/validation");
 
+const { createTemplate } = require("./template/resolvers");
+
 const gqlClient = new GraphQLClient(process.env.HASURA_GRAPHQL_URL, {
   headers: {
     "x-hasura-admin-secret": process.env.HASURA_ADMIN_SECRET,
   },
 });
-
-const getUsersTemplateGroupSharedTo = gql`
-  query MyQuery($groupId: uuid!) {
-    shared_template_groups(where: { template_group_id: { _eq: $groupId } }) {
-      shared_to_user {
-        id
-        name
-        image
-        email
-      }
-    }
-  }
-`;
 
 const shareTemplateQuery = gql`
   mutation shareTemplate(
@@ -96,84 +85,7 @@ const resolvers = {
 
       return res?.update_api_keys?.returning?.[0];
     },
-    insert_template: async (_, args, { userId }) => {
-      if (!userId) return;
-
-      const mutation = gql`
-        mutation (
-          $name: String!
-          $prompts: jsonb
-          $files: jsonb!
-          $template_group_id: uuid
-          $owner_id: uuid!
-        ) {
-          insert_templates_one(
-            object: {
-              name: $name
-              files: $files
-              prompts: $prompts
-              template_group_id: $template_group_id
-              owner_id: $owner_id
-            }
-          ) {
-            id
-            name
-            prompts
-            files
-            template_group_id
-            owner_id
-          }
-        }
-      `;
-
-      const { name, prompts, files, template_group_id } = args.object;
-
-      const isPromptsValid = validatePrompts(JSON.parse(prompts));
-      const isFilesValid = validateFiles(JSON.parse(files));
-
-      if (!isPromptsValid) {
-        throw new Error("Prompts not valid");
-      }
-
-      if (!isFilesValid) {
-        throw new Error("Files not valid");
-      }
-
-      const res = await gqlClient.request(mutation, {
-        name,
-        prompts,
-        files,
-        template_group_id,
-        owner_id: userId,
-      });
-
-      // if group was shared - share this template with all that users
-      if (template_group_id && res?.insert_templates_one?.id) {
-        // find users group was shared to
-        const users = await gqlClient.request(getUsersTemplateGroupSharedTo, {
-          groupId: template_group_id,
-        });
-        const userEmails =
-          users?.shared_template_groups.map(
-            (item) => item.shared_to_user.email
-          ) || [];
-
-        // share template
-        if (userEmails.length > 0) {
-          await Promise.all(
-            userEmails.map((email) =>
-              gqlClient.request(shareTemplateQuery, {
-                templateId: res.insert_templates_one.id,
-                shareToUserEmail: email,
-                shareByUserId: userId,
-              })
-            )
-          );
-        }
-      }
-
-      return res?.insert_templates_one;
-    },
+    insert_template: createTemplate,
     update_template: async (_, args, { userId }) => {
       if (!userId) return;
